Reduce the Epoll Events array to just one element
It appears that there is some marshalling between Ada and C that is resulting in
the boundaries of this array of pointers to not work properly. With larger array
sizes, whenever there are concurrent connections to Mozzoni it results in
invalid Polled_Event.Data.FD values.

This is a **hack** right now, and the underlying bug still remains at large.