Refactor and rework the epoll utilization in the main subprogram
The primary issue addressed here is the creation of a new stack-allocated
Event_Type whenever we're added a socket to the epoll file descriptor. The
behavior previously was re-using the same object, and I _believe_ resulting in
only one socket (last one wins) being monitored.

With this change we're properly handling concurrent connections